Add files via upload
I changed quite a bit and ended with a faulty code because I did not know some of Bootstrap's best practices.

I used too many position: relative and because there are objects, which are hidden and allocate space, there is a problem with dropping container after container. Hence, I used negative values and position: relative.

The design is still very responsive, though it can improve greatly.